Refactor : 장소 업로드 - 위치 선택자 부분 수정중
위치 선택자 부분 Refactoring part one
