+++ client/project-there/src/components/placeUpload/placeLocationSelector/PlaceLocationSelectorMap.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function PlaceLocationSelectorMap() {
+
+}
