+++ client/project-there/src/components/placeUpload/placeLocationSelector/PlaceSearchResultElement.jsx
