+++ client/project-there/src/components/placeUpload/placeLocationSelector/PlaceSearchResultList.jsx
